pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dados-br"
version = "0.2.0"
description = "Ferramenta CLI open source para catalogar, baixar, validar e organizar dados públicos brasileiros"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.11"
authors = [{ name = "dados-br Contributors" }]
keywords = ["dados-abertos", "brasil", "data", "opendata", "governo", "inep", "ibge", "datasus", "transparencia", "ministerio-publico", "prefeituras", "capes", "cnpq"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Utilities",
]
dependencies = [
    "typer>=0.12.0",
    "rich>=13.7.0",
    "pydantic>=2.6.0",
    "PyYAML>=6.0.1",
    "httpx>=0.27.0",
    "pandas>=2.2.0",
    "pyarrow>=15.0.0",
    "duckdb>=0.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.0",
    "pytest-cov>=5.0.0",
    "respx>=0.21.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "types-PyYAML>=6.0.12",
]

[project.scripts]
dados-br = "dadosbr.cli:app"

[project.urls]
Homepage = "https://github.com/dados-br/dados-br"
Documentation = "https://github.com/dados-br/dados-br#readme"
Issues = "https://github.com/dados-br/dados-br/issues"
Changelog = "https://github.com/dados-br/dados-br/blob/main/CHANGELOG.md"

[tool.hatch.build.targets.wheel]
packages = ["dadosbr"]

[tool.hatch.build.targets.sdist]
include = ["dadosbr/", "catalog/", "tests/", "README.md", "LICENSE", "CONTRIBUTING.md"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short --strict-markers"
markers = ["integration: marca testes que fazem requisições reais à internet"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.coverage.run]
source = ["dadosbr"]
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true
skip_covered = false
